[D8] Fixes contact block. Use marketo field ids for field mappings.

+++ src/MarketoMaService.php
@@ -2,7 +2,6 @@
 
 namespace Drupal\marketo_ma;
 
-use Drupal\Component\Utility\NestedArray;
 use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Core\Path\PathMatcherInterface;
 use Drupal\Core\Queue\QueueFactory;
@@ -142,24 +141,36 @@ class MarketoMaService implements MarketoMaServiceInterface {
       }
 
       // Get the Lead data from temporary user storage.
-      $lead = new Lead($this->getUserData());
+      if (! $lead = $this->getUserData()) {
+        $lead = new Lead();
+      }
+
       // Check for logged in user.
       if (empty($lead->getEmail()) && !empty($this->current_user->getEmail())) {
         $lead->set('email', $this->current_user->getEmail());
       }
+
       // Check for the munchkin option and that the munchkin api is configured.
-      if ($this->trackingMethod() == 'munchkin' && $this->munchkin->isConfigured() && !empty($lead->getEmail()) && $lead->get('associated') !== TRUE) {
+      if ($this->trackingMethod() == MarketoMaServiceInterface::TRACKING_METHOD_MUNCHKIN
+        && $this->munchkin->isConfigured()
+        && !empty($lead->getEmail())
+        && $lead->get('associated') !== TRUE
+      ) {
         // Set drupalSettings so JS will do the lead association.
         $page['#attached']['drupalSettings']['marketo_ma']['actions'][] = $this->munchkin->getAction(MarketoMaMunchkinInterface::ACTION_ASSOCIATE_LEAD, $lead);
         // Set the associated flag so we are not associating on every request.
-        $this->setUserData($lead->data() + ['associated' => TRUE]);
+        $this->setUserData($lead->set('associated', TRUE));
       }
       // Check for the api option and that the client can connect.
-      elseif ($this->trackingMethod() == 'api_client' && $this->api_client->canConnect() && !empty($lead->getEmail()) && $lead->get('associated') !== TRUE) {
+      elseif ($this->trackingMethod() == MarketoMaServiceInterface::TRACKING_METHOD_API
+        && $this->api_client->canConnect()
+        && !empty($lead->getEmail())
+        && $lead->get('associated') !== TRUE
+      ) {
         // Use the API to associate the lead.
         $this->updateLead($lead);
         // Set the associated flag so we are not associating on every request.
-        $this->setUserData($lead->data() + ['associated' => TRUE]);
+        $this->setUserData($lead->set('associated', TRUE));
       }
     }
   }
@@ -204,7 +215,7 @@ class MarketoMaService implements MarketoMaServiceInterface {
    */
   public function updateLead($lead) {
     // Get the tracking method.
-    if ($this->trackingMethod() === 'api_client') {
+    if ($this->trackingMethod() === MarketoMaServiceInterface::TRACKING_METHOD_API) {
       // Do we need to batch the lead update?
       if (!$this->config()->get('rest.batch_requests')) {
         // Just sync the lead now.
@@ -213,13 +224,11 @@ class MarketoMaService implements MarketoMaServiceInterface {
         // Queue up the lead sync.
         $this->queue_factory->get('marketo_ma_lead')->createItem($lead);
       }
-      // Save the lead data in the user data for future use.
-      if (!empty($this->current_user->getLastAccessedTime())) {
-        $this->setUserData(NestedArray::mergeDeep($lead->data(), !empty($this->getUserData()) ? $this->getUserData() : []));
-      }
+
+      $this->resetUserData();
     } else {
-      // Save the data for the munchkin API.
-      $this->setUserData($lead->data());
+      // Set the user data so munchkin can take it from there.
+      $this->setUserData($lead);
     }
 
     return $this;
@@ -228,8 +237,11 @@ class MarketoMaService implements MarketoMaServiceInterface {
   /**
    * {@inheritdoc}
    */
-  public function setUserData($data) {
-    $this->temporaryStorage()->set('user_data', $data);
+  public function setUserData($lead) {
+    // Make sure we have a real user before trying to access user data.
+    if (!empty($this->current_user->getLastAccessedTime())) {
+      $this->temporaryStorage()->set('user_data', $lead);
+    }
     return $this;
   }
 
@@ -244,7 +256,10 @@ class MarketoMaService implements MarketoMaServiceInterface {
    * {@inheritdoc}
    */
   public function resetUserData() {
-    $this->temporaryStorage()->delete('user_data');
+    // Make sure we have a real user before trying to access user data.
+    if (!empty($this->current_user->getLastAccessedTime())) {
+      $this->temporaryStorage()->delete('user_data');
+    }
     return $this;
   }
 
@@ -258,30 +273,43 @@ class MarketoMaService implements MarketoMaServiceInterface {
   /**
    * {@inheritdoc}
    */
-  public function getAvailableFields($reset = FALSE) {
+  public function getMarketoFields($reset = FALSE) {
     // Reset if requested or fields have never been retrieved.
     if ($reset || $this->state->get('marketo_ma.field.defined_fields', FALSE) === FALSE) {
       // Get the fields.
       $api_fields = $this->api_client->canConnect() ? $this->api_client->getFields() : [];
 
-      // Loop through and sanitize the values.
-      $field_options = [];
-      foreach ($api_fields as $row) {
-        $field_options[$row['id']] = [
-          $this->t(':value', [':value' => !isset($row['id']) ? '' : $row['id']]),
-          $this->t(':value', [':value' => !isset($row['displayName']) ? '' : $row['displayName']]),
-          $this->t(':value', [':value' => !isset($row['rest']['name']) ? '' : $row['rest']['name']]),
-          $this->t(':value', [':value' => !isset($row['soap']['name']) ? '' : $row['soap']['name']]),
-        ];
-      }
-
       // Save the field options in state.
-      if (!empty($field_options)) {
-        $this->state->set('marketo_ma.field.defined_fields', $field_options);
+      if (!empty($api_fields)) {
+        $marketo_ma_fields = [];
+        // Convert response array to an array of MarketoField objects keyed by the marketo field id.
+        foreach ($api_fields as $api_field) {
+          $marketo_ma_fields[$api_field['id']] = new MarketoFieldDefinition($api_field);
+        }
+        $this->state->set('marketo_ma.field.defined_fields', $marketo_ma_fields);
       }
     }
 
     return $this->state->get('marketo_ma.field.defined_fields', []);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getMarketoFieldsAsTableSelectOptions($reset = FALSE) {
+    // Convert objects to table-select options.
+    return array_map(function ($item) {
+      return $item instanceof MarketoFieldDefinition
+        ? $item->toTableSelectOption()
+        : [];
+    }, $this->getMarketoFields($reset));
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getEnabledFields() {
+    return array_intersect_key($this->getMarketoFields(), (array) $this->config()->get('field.enabled_fields'));
   }
 
   /**
